updated spec file path in congid and moved test files to specs dir
